requirements.txt
datasets
h5py~=3.13.0
joblib~=1.4.2
librosa~=0.10.2.post1
matplotlib~=3.10.1
numpy~=2.1.3
pandas~=2.2.3
phonemizer
pydub~=0.25.1
pytubefix~=8.12.2
PyYAML~=6.0.2
requests~=2.32.3
scikit-learn~=1.6.1
seaborn
spacy
soundfile~=0.13.1
transformers~=4.50.0
torch~=2.6.0
tqdm~=4.67.1
whisperx~=3.3.1